pyproject.toml
[project]
name = "crabot"
version = "0.1.0"
description = "random basic discord bot"
requires-python = ">=3.14"
dependencies = [
    "aiocache>=0.12.3",
    "aiohttp>=3.13.3",
    "asyncio>=4.0.0",
    "cachetools>=7.0.5",
    "google-adk>=1.28.1",
    "langfuse>=4.0.6",
    "openinference-instrumentation-google-adk>=0.1.10",
    "pydantic>=2.12.5",
]
